jandinote/requirements.txt
asgiref==3.6.0
Django==4.2
python-environ==0.4.54
sqlparse==0.4.4
